sales-and-inventory-management/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
defusedxml==0.7.1
Django==4.2.6
django-appconf==1.0.5
django-autoslug==1.9.9
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.3
django-imagekit==5.0.0
django-phonenumber-field==7.2.0
django-tables2==2.6.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.26.0
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.1
phonenumbers==8.13.22
pilkit==3.0
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tabulate==0.9.0
timeago==1.0.16
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
xlrd==2.0.1
xlwt==1.3.0

qrcode~=7.4.2
opencv-python~=4.8.1.78